requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
attrs==23.1.0
    # via
    #   automat
    #   outcome
    #   service-identity
    #   trio
    #   twisted
automat==22.10.0
    # via twisted
beautifulsoup4==4.12.2
    # via -r requirements.in
certifi==2023.11.17
    # via
    #   requests
    #   selenium
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
constantly==23.10.4
    # via twisted
cryptography==41.0.5
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.2.0
    # via
    #   parsel
    #   scrapy
fastapi==0.104.0
    # via -r requirements.in
filelock==3.13.1
    # via tldextract
greenlet==3.0.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   uvicorn
    #   wsproto
hyperlink==21.0.0
    # via twisted
idna==3.4
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   tldextract
    #   trio
incremental==22.10.0
    # via twisted
itemadapter==0.8.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.1.0
    # via scrapy
jinja2==3.1.2
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   itemloaders
    #   parsel
lxml==4.9.3
    # via
    #   parsel
    #   scrapy
markupsafe==2.1.3
    # via jinja2
outcome==1.3.0.post0
    # via trio
packaging==23.2
    # via
    #   parsel
    #   scrapy
parsel==1.8.1
    # via
    #   itemloaders
    #   scrapy
protego==0.3.0
    # via scrapy
psycopg[binary]==3.1.13
    # via -r requirements.in
psycopg-binary==3.1.13
    # via psycopg
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pydantic==2.5.1
    # via fastapi
pydantic-core==2.14.3
    # via pydantic
pydispatcher==2.0.7
    # via scrapy
pyopenssl==23.3.0
    # via scrapy
pysocks==1.7.1
    # via urllib3
pyyaml==6.0.1
    # via -r requirements.in
queuelib==1.6.2
    # via scrapy
requests==2.31.0
    # via
    #   requests-file
    #   tldextract
requests-file==1.5.1
    # via tldextract
scrapy==2.11.0
    # via -r requirements.in
selenium==4.15.2
    # via -r requirements.in
service-identity==23.1.0
    # via scrapy
six==1.16.0
    # via
    #   automat
    #   requests-file
sniffio==1.3.0
    # via
    #   anyio
    #   trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.22
    # via
    #   -r requirements.in
    #   sqlalchemy-enum34
sqlalchemy-enum34==2.0.0
    # via -r requirements.in
starlette==0.27.0
    # via fastapi
tldextract==5.1.1
    # via scrapy
trio==0.23.1
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
twisted==22.10.0
    # via scrapy
typing-extensions==4.8.0
    # via
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   twisted
urllib3[socks]==2.1.0
    # via
    #   requests
    #   selenium
uvicorn==0.24.0
    # via -r requirements.in
w3lib==2.1.2
    # via
    #   itemloaders
    #   parsel
    #   scrapy
wsproto==1.2.0
    # via trio-websocket
zope-interface==6.1
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
